requirements.txt
PyJWT==2.9.0
cryptography==43.0.1
requests==2.32.3
PyMySQL==1.1.1